remove unused operator
